poptorch_geometric/requirements.txt
# IMPORTANT: Keep requirements in sync with config.buildenv.py

# Install pre-built wheels for PyTorch Geometric that are compatible with
# poptorch which is currently pinned to torch 1.13.1
--find-links https://data.pyg.org/whl/torch-1.13.1+cpu.html

pyg-nightly==2.2.0.dev20230110

pytest-benchmark
pytest-cov
torch-scatter
torch-sparse
torch-cluster
torch-spline-conv

nbformat
nbconvert
